+++ apps/asterum_traveler-app/babel.config.cjs
@@ -8,5 +8,15 @@ module.exports = {
     '@babel/plugin-transform-runtime',
     'babel-plugin-transform-import-meta',
     'babel-plugin-transform-vite-meta-env',
+    'babel-plugin-styled-components',
+    [
+      'module-resolver',
+      {
+        root: ['./src'],
+        alias: {
+          '@': './src',
+        },
+      },
+    ],
   ],
 };