+++ apps/asterum_traveler-app/CHANGELOG.md
@@ -2,6 +2,31 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [2.0.0](https://github.com/clappingmin/asterum_traveler/compare/app-v1.1.0...app-v2.0.0) (2025-03-17)
+
+
+### Bug Fixes
+
+* 요청 서버로 보내기 ([bbc38ea](https://github.com/clappingmin/asterum_traveler/commit/bbc38ea67b0db324bb8a29b68f864ff973feb12d))
+* 요청 서버로 보내기 req url 수정 ([ea2e4a2](https://github.com/clappingmin/asterum_traveler/commit/ea2e4a2c6c66b855bb8070736e5e2c8b92774a63))
+* build 오류 test코드 임포트 오류 수정 ([856762c](https://github.com/clappingmin/asterum_traveler/commit/856762cf148598ebbf5c29903acbad5e90b7a064))
+* index.html 수정 ([dd39cb3](https://github.com/clappingmin/asterum_traveler/commit/dd39cb33251d4af51c71931a2e05ac8f957bdf50))
+* queryClient를 main이 아니라 renderer/Pageshell에 선언한거 사용하기 ([8569d87](https://github.com/clappingmin/asterum_traveler/commit/8569d87362c17f64fd29de8b45f993e19a0bd150))
+* ts를 js로 변경 ([fba5164](https://github.com/clappingmin/asterum_traveler/commit/fba516464cc16e15f8bc56bfcd5d9c76fcce5653))
+* ts를 js로 변경 ([0f64be2](https://github.com/clappingmin/asterum_traveler/commit/0f64be258f8bb86f82d524710bf758e9067383b5))
+* Uncaught Error: [vite-plugin-ssr][Wrong Usage] pageContext["urlPathname"] isn't available on the client-side because "urlPathname" is missing in passToClient, see https://vite-plugin-ssr.com/passToClient ([065b4c8](https://github.com/clappingmin/asterum_traveler/commit/065b4c89df415bc967cd7edfbd6f8f870e56412b))
+* vercel 빌드 오류 수정 ([7408b61](https://github.com/clappingmin/asterum_traveler/commit/7408b6158ef337ccfbd0fcf0cfcdb8ee352e9819))
+* vercel 빌드 오류 수정 ([028a7ee](https://github.com/clappingmin/asterum_traveler/commit/028a7eed1b16a1626f41dfb7a625106622b7766c))
+* vercel 빌드 오류 수정 ([1a300f6](https://github.com/clappingmin/asterum_traveler/commit/1a300f68674cd499e270f9a0b4d6f8fae87a0c52))
+* vercel 빌드 오류 수정 ([b25932f](https://github.com/clappingmin/asterum_traveler/commit/b25932facef28bb12aedbf59ce1d019a4b9cf813))
+* vercel 서버 설정 ([7354527](https://github.com/clappingmin/asterum_traveler/commit/7354527fd948ba265de4dd5d2ce0dbb17b8a36a7))
+* vercel 서버 설정 ([342a225](https://github.com/clappingmin/asterum_traveler/commit/342a225b33cb3c9b9c08ba590ab99477fdbadb8a))
+* vercel 서버 설정 ([f64448e](https://github.com/clappingmin/asterum_traveler/commit/f64448e4c7b1dbfc02a867e54219dabc3c1e9586))
+* vercel notfound 에러 ([8fce793](https://github.com/clappingmin/asterum_traveler/commit/8fce793119a088a54ede221f944baa718c9711e0))
+* vercel ssr 생성 및 알리기 완료 ([dfc7fc2](https://github.com/clappingmin/asterum_traveler/commit/dfc7fc2eb8a943ea66c1e3c0cad17c06a7f4ab48))
+* vercel ssr 서버 알리는 코드 수정 ([69efc52](https://github.com/clappingmin/asterum_traveler/commit/69efc520b440e7bf689a8dc519573678901ee975))
+* vite-plugin-ssr 버전 픽스 ([46fc859](https://github.com/clappingmin/asterum_traveler/commit/46fc859f364a45099480499e174fcf54c02add66))
+
 ## [2.0.0](https://github.com/clappingmin/asterum_traveler/compare/app-v2.0.0...app-v2.0.0) (2025-03-12)
 
 ## [2.0.0](https://github.com/clappingmin/asterum_traveler/compare/app-v1.0.2...app-v2.0.0) (2025-03-12)